+++ src/Http/Middleware/CsrfProtectionMiddleware.php
@@ -23,14 +23,11 @@ class CsrfProtectionMiddleware extends Middleware
 {
     protected $defaultConfig = [
         'cookieName' => 'CSRF-Token', // Cookie name for generated CSRF token
-        'expires' => '+60 minutes', // Expiry time defaults to 60 minutes like sessions
+        'expires' => '+60 minutes', // Expiry time defaults to 60 minutes like sessions,
+        'httpOnly' => false, // only available to HTTP protocol (not javascript)
+        'secure' => false, // cookie only transmitted over HTTPS,
+        'encrypt' => true
     ];
-
-    /**
-    * Internal flag
-    * @var boolean
-    */
-    private $createCookie = false;
 
     /**
      * This handles the request
@@ -46,7 +43,7 @@ class CsrfProtectionMiddleware extends Middleware
         if ($request->params('csrfProtection') === false) {
             return ;
         }
-        
+  
         # Generate the CSRF token
         $token = $request->cookies($this->config('cookieName'));
         if ($request->is(['get']) and $token === null) {
@@ -78,9 +75,9 @@ class CsrfProtectionMiddleware extends Middleware
         if ($request->params('csrfProtection') === false) {
             return ;
         }
-
-        if ($request->is(['get']) and $this->createCookie) {
-            $response->cookie($this->config('cookieName'), $request->params('csrfToken'), ['expires' => $this->config('expires')]);
+   
+        if ($request->method() === 'GET') {
+            $response->cookie($this->config('cookieName'), $request->params('csrfToken'), $this->config);
         }
     }
 
@@ -127,7 +124,6 @@ class CsrfProtectionMiddleware extends Middleware
     */
     private function generateToken() : string
     {
-        $this->createCookie = true;
         $randomBytes = random_bytes(64);
 
         return Security::hash($randomBytes, [