+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 ## [Unreleased]
+
+## [2.0.1] - 2019-10-30
+### Fixed
+- OriginTestCase changed to Abstract class which was causing unexpected issues with PHPUnit
+- Fixed CSRF token now rewrites cookie with each GET request
+- Fixed loadMiddleware accepts options
+- Paginator behavior with pages that are out of bounds throws page not found instead of ignoring
+
 ## [2.0.0] - 2019-10-25
 
 Releasing version 2.0, see the [upgrade guide](https://www.originphp.com/docs/upgrade) for information on how to upgrade from previous versions.